+++ dashboard/src/main.ts
@@ -28,8 +28,8 @@ import ModuleParametersView from "@/components/model/ModuleParameters.vue";
 import TemperatureParametersView from "@/components/model/TemperatureParameters.vue";
 import TrackingParametersView from "@/components/model/TrackingParameters.vue";
 import SystemView from "@/components/model/System.vue";
-import WeatherUpload from "@/components/jobs/WeatherUpload.vue";
-import WeatherCSVMapper from "@/components/jobs/WeatherCSVMapper.vue";
+import WeatherUpload from "@/components/jobs_dev/WeatherUpload.vue";
+import WeatherCSVMapper from "@/components/jobs_dev/WeatherCSVMapper.vue";
 
 import "./assets/css/styles.css";
